finish cleaning up mingw project for moving to a single mingw-w64 compiler that compiles BOTH 32 and 64 bit executables

+++ src/Cpl/Io/File/_0test/readwrite.cpp
@@ -56,7 +56,7 @@ TEST_CASE( "readwrite", "[readwrite]" )
     REQUIRE( fd.write( myBuffer, sizeof(myBuffer) ) );
     for(size_t i=0; i<sizeof(myBuffer); i++) { sum += myBuffer[i]; }
     REQUIRE( fd.write( myBuffer, sizeof(myBuffer), bytesWritten ) );
-    REQUIRE( bytesWritten == sizeof(myBuffer) );
+    REQUIRE( (size_t)bytesWritten == sizeof(myBuffer) );                                                                            
     for(int i=0; i<bytesWritten; i++) { sum += myBuffer[i]; }
     fd.flush(); 
 
@@ -104,5 +104,5 @@ TEST_CASE( "readwrite", "[readwrite]" )
     REQUIRE( writer.println() == false );
     REQUIRE( fd2.write('a') == false );
 
-    REQUIRE( Cpl::System::Shutdown_TS::getAndClearCounter() == 0 );
+    REQUIRE( Cpl::System::Shutdown_TS::getAndClearCounter() == 0u );
     }
